requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
httpx
aiogram
pydantic
apscheduler
requests
python-dotenv
alembic
aiohttp
asyncio